+++ src/qc_model.jl
@@ -44,6 +44,7 @@ function variable_QCModel_balas(qcm::QuantumCircuitModel)
 
     if qcm.data["decomposition_type"] == "approximate"
         QCO.variable_slack_for_feasibility(qcm)
+        QCO.variable_slack_var_outer_approximation(qcm)
     end
 
     QCO.variable_QCModel_valid(qcm)
@@ -60,6 +61,7 @@ function constraint_QCModel_balas(qcm::QuantumCircuitModel)
     QCO.constraint_gate_product_linearization(qcm)
     QCO.constraint_gate_target_condition(qcm)
     QCO.constraint_cnot_gate_bounds(qcm)
+    (qcm.data["decomposition_type"] == "approximate") && (QCO.constraint_slack_var_outer_approximation(qcm))
     (!qcm.data["are_gates_real"]) && (QCO.constraint_complex_to_real_symmetry(qcm))
 
     QCO.constraint_QCModel_valid(qcm)
@@ -75,6 +77,7 @@ function variable_QCModel_compact(qcm::QuantumCircuitModel)
 
     if qcm.data["decomposition_type"] == "approximate"
         QCO.variable_slack_for_feasibility(qcm)
+        QCO.variable_slack_var_outer_approximation(qcm)
     end
 
     QCO.variable_QCModel_valid(qcm)
@@ -106,6 +109,7 @@ function constraint_QCModel_compact(qcm::QuantumCircuitModel)
     QCO.constraint_gate_product_linearization(qcm)
     QCO.constraint_gate_target_condition_compact(qcm)
     QCO.constraint_cnot_gate_bounds(qcm)
+    (qcm.data["decomposition_type"] == "approximate") && (QCO.constraint_slack_var_outer_approximation(qcm))
     # (!qcm.data["are_gates_real"]) && (QCO.constraint_complex_to_real_symmetry(qcm))
 
     QCO.constraint_QCModel_valid(qcm)